requirements.txt
Django==1.4.2
South==0.7.6
amqplib==1.0.2
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.18
celery==3.0.12
celery-with-redis==3.0
coverage==3.5.3
django-appconf==0.5
django-celery==3.0.11
django-compressor==1.2
django-debug-toolbar==0.9.4
django-extensions==1.0.1
django-nose==1.1
eventlet==0.9.17
gevent==0.13.8
gevent-socketio==0.3.5-rc2
gevent-websocket==0.3.6
greenlet==0.4.0
gunicorn==0.15.0
kombu==2.4.8
mock==1.0.1
nose==1.2.1
python-dateutil==1.5
pytz==2012h
redis==2.7.1
wsgiref==0.1.2
